fix: cell tower area is u16

+++ src/model.rs
@@ -9,7 +9,7 @@ pub enum Transmitter {
         radio: CellRadio,
         country: u16,
         network: u16,
-        area: u32,
+        area: u16,
         cell: u64,
         unit: u16,
     },
